requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langgraph==0.6.6
pydantic==2.11.9
requests==2.32.5
pillow==11.3.0
huggingface_hub==0.34.4
python-dotenv==1.1.1
